requirements.txt
mysql-connector-python
psycopg2-binary
Faker
python-dotenv
sqlalchemy
pandas